Include all custom peripherals in robot test harness (#4)
* Include all custom peripherals in robot test harness

OTPMemory, I2CFaultProxy, STM32F4FlashInterceptor, and
STM32H7FlashController were added after the original peripheral set
but never included in ota_fault_point.robot. This caused OTP security
profiles to fail in CI with exit=2 since Renode couldn't resolve the
Memory.OTPMemory type from the platform description.

* Include all custom peripherals in robot harness, fix I2C compile error

Add OTPMemory, I2CFaultProxy, STM32F4FlashInterceptor, and
STM32H7FlashController to ota_fault_point.robot default includes.
Fix CS0034 ambiguous operator in I2CFaultProxy.cs line 463 (int * ulong).

* Wire OTP, NVS, I2C, instruction-skip, and read-fault variables through robot

13 resc variables for newer fault families were never declared or passed
in ota_fault_point.robot. OTP profiles silently got empty otp_peripheral,
causing calibration to fail with exit=2 in CI.

* Route OTP fault types through separate OTP peripheral, print stderr on failure

OTP profiles use flash_backend=nvm_ctrl (kind='slow') with a separate
otp_peripheral. OTP fault codes (op/os/od/oo) were dispatching through
arm_fault() which targets the flash backend, never reaching the OTP
device. Added arm_otp_fault/disarm_otp_fault/was_otp_fault_injected
helpers and an is_otp_fault branch in run_execute_fault.

Also print last 20 lines of stderr in self_test.py when a profile
fails with no report, so CI failures aren't blind.

* Skip OTP fault points when calibration finds 0 OTP blows

OTP profiles were timing out because the fault plan generated OTP fault
points using flash write indices (firmware doesn't blow OTP fuses).
Mirror the I2C pattern: emit total_otp_blows from calibration, propagate
through CalibrationResult/CalibrationInputs, and guard OTP fault point
generation on total_otp_blows > 0.

+++ peripherals/I2CFaultProxy.cs
@@ -460,7 +460,7 @@ namespace Antmicro.Renode.Peripherals.I2C
             Array.Copy(data, corrupted, data.Length);
 
             uint seed = faultSeed != 0 ? faultSeed : (uint)totalTransactions;
-            seed ^= (uint)(data.Length * 2654435761UL);
+            seed ^= (uint)((ulong)data.Length * 2654435761UL);
 
             // Flip 1-4 bits across the data.
             int flips = 1 + (int)(NextLcg(ref seed) % 4U);
